ci: add GitHub Actions workflow (fmt + clippy + test + doc) + clean up debt (#31)
ferrolearn had no PR CI before this — only Dependabot at the org level.
This adds a strict, parallel-jobs GitHub Actions workflow plus the fmt
+ clippy cleanup needed to land it green from day one.

## New workflow (.github/workflows/ci.yml)

Four parallel jobs on every push to main + every PR targeting main:

- `cargo fmt --check`           — strict rustfmt parity with main
- `cargo clippy -D warnings`    — strict lints across `--workspace --all-targets`
- `cargo test --release`        — full workspace test suite in release
- `cargo doc --no-deps`         — docs build (no `-D warnings` yet; see below)

Architectural choices:

- **Independent jobs**: fmt / clippy / test / doc all run in parallel so
  one failure doesn't gate the others — contributors see the full picture
  in a single CI run.
- **Concurrency control**: `cancel-in-progress` on the same ref so a
  fix-up push kills the stale run.
- **Permissions hardening**: `contents: read` only — workflow doesn't
  push, comment, or deploy.
- **Caching**: `Swatinem/rust-cache@v2` keyed per-job so each artifact
  pool stays isolated (clippy vs test vs doc).
- **Stable only**: no MSRV matrix yet — workspace declares `rust-version
  = "1.85"` but enforcement can come as a separate matrix job if MSRV
  regressions become a practical concern.
- **Doc job is advisory**: `RUSTDOCFLAGS=-D warnings` would be ideal but
  the workspace currently has multiple broken intra-doc links
  (`FerroError::*` variants in `ferrolearn-io`, etc.). Job catches
  syntax-level doc errors that would break the docs.rs build; lint
  tightening tracked as follow-up.

## Pre-CI cleanup so it lands green

- **`cargo fmt` applied across the workspace** (66 files, ~1500 line
  diff, all whitespace). Pure mechanical churn; brings the tree to a
  rustfmt-clean baseline so future drift gets caught immediately.

- **9 clippy violations fixed** to clear `-D warnings`:
  - 8× `clippy::needless_range_loop` (6 files: `categorical.rs`,
    `gmm.rs`, `lda.rs`, `gradient_boosting.rs`, `hist_gradient_boosting.rs`,
    `cluster/tests/conformance_sklearn.rs`). Each fixed with a
    scoped `#[allow(clippy::needless_range_loop)]` on the enclosing
    function — these are math-style matrix index loops (per-class score
    rows, smoothing, k-means++ candidate search) where range-by-index
    is the natural form; rewriting to `.iter().enumerate()` would
    obscure the intent. The allow is explicit and per-function so
    new loops in adjacent code still get linted.
  - 1× `clippy::approximate_constant` — false positive in
    `ferrolearn-neural/src/mlp.rs` test: `3.14` was a placeholder for
    the Identity activation passthrough test, not an approximation of
    π. Changed to `1.25` with a comment.
  - 1× `clippy::unused_imports` — `Transform` not used in
    `ferrolearn-neural/tests/conformance_wave4.rs`.

- **Auto-fixed by `cargo clippy --fix`**: several `clippy::redundant_closure`
  / `clippy::derefed_type_is_same_as_origin` / `clippy::manual_range_patterns`
  / `clippy::length_comparison_to_one` issues across
  `ferrolearn-preprocess`, `ferrolearn-tree`, `ferrolearn-linear`,
  `ferrolearn-numerical`, `ferrolearn-model-sel`. ~10 files touched.

## Verification

```
cargo fmt --check                                                    # clean
cargo clippy --workspace --all-targets -- -D warnings                # clean
cargo test --workspace --release --no-run                            # builds
cargo doc --workspace --no-deps --document-private-items             # builds (1 warning)
```

## Known follow-ups (out of scope here)

1. Doc-link cleanup: ~10 broken intra-doc links in `ferrolearn-io` +
   `ferrolearn-bayes` that prevent tightening to `RUSTDOCFLAGS=-D warnings`.
2. MSRV enforcement: add a `cargo +1.85 check` matrix job if MSRV
   regressions show up.
3. Coverage / dependabot review: not added; can come later.

Refs forecast-bio/decode#3 (upstream contributions backlog).

+++ ferrolearn-preprocess/tests/api_proof.rs
@@ -273,7 +273,7 @@ fn api_proof_text() {
     let f = CountVectorizer::new().fit(&docs).unwrap();
     let counts = f.transform(&docs).unwrap();
     assert_eq!(counts.nrows(), 3);
-    let counts_f64 = counts.mapv(|v| v as f64);
+    let counts_f64 = counts.mapv(|v| v);
     let f = TfidfTransformer::<f64>::new().fit(&counts_f64).unwrap();
     let _ = f.transform(&counts_f64).unwrap();
 }
